+++ src/util/mod.rs
@@ -1,0 +1,4 @@
+// Copyright (C) 2026 gabkhanfig
+// SPDX-License-Identifier: GPL-2.0-only
+
+pub mod bin_version;